backend/requirements.txt
alembic==1.14.1
alicat==0.8.0
annotated-types==0.7.0
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.55.3
future==1.0.0
greenlet==3.1.1
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
SciencePlots==2.1.1
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
